refactor(desktop): split mesh serving usage
